requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
requests==2.31.0
cryptography==41.0.7
Pillow==10.0.1
gunicorn==22.0.0
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.5.0
whitenoise==6.7.0